remove amount from tests
Amount is not used for anything, so better to remove it.
To ensure determinism with the tests, the addresses are now sorted lexicographically
before generating outputs. This is necessary because the outputs will change based
on the order in which addresses for the same recipient are created.

For the receiver, this doesnt matter. This is only relevant for determinism in the
test vectors.